backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
celery>=5.3.6
redis>=5.0.1
psycopg2-binary>=2.9.9
requests>=2.31.0
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
cloudinary>=1.36.0
Pillow>=10.0.0
gunicorn>=21.2.0
pytest>=8.0.0
pytest-django>=4.8.0